requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.3
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-betterforms==1.2
django-ckeditor==5.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pilkit==2.0
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.1
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
